requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
redis>=5.0.4
httpx>=0.27.0
anthropic>=0.28.0
langgraph>=0.1.0
python-multipart>=0.0.9
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0